# Conflicts:
#	TowerGameJS-Phase4/Thumbs.db

+++ TowerGameJS-Phase4/game.js
@@ -122,7 +122,7 @@ class Game {
     // An adjacent neighbor has a step of 10
     // and a diagonal neighbor has a step of 14.
 
-  brushfire() {
+  brushfire(undo) { //if brushfire fails to make a valid map undo will be called
     // Initialize each cell in the grid to have a distance that
     // is the greatest possible.  Initialize each cell to
     // have no parent and populate it's array of neighbors
@@ -162,13 +162,22 @@ class Game {
                 }
           }     // for each neighbor
         }   // while(queue.length)
+    if(!this.validMap()){
+      if(undo){
+          undo()
+          this.brushfire()
+      }else{
+        // delete any enemy that is currently in a cell without a parent
+        for(let i = 0; i < this.enemies.length;  i++) {
+            let enemy = towerGame.enemies[i];
+            if(!enemy.currentCell.parent)
+                enemy.kill = true;    // kill the orphans
+            }
+            console.log("brushfire created an invalid map and no undo was inputed")
 
-    // delete any enemy that is currently in a cell without a parent
-    for(let i = 0; i < this.enemies.length;  i++) {
-        let enemy = towerGame.enemies[i];
-        if(!enemy.currentCell.parent)
-            enemy.kill = true;    // kill the orphans
-        }
+      }
+    }
+
 
         // give each cell a vector that points to its parent
 //       for(var i = 0; i < this.cols; i++){
@@ -178,7 +187,34 @@ class Game {
 //       }
 
     }
+    //check the map to see if there are cells without parents
+    validMap() {
+      for(var i = 0; i < this.cols; i++){
+        for(var j = 0; j < this.rows; j++){
+          var cell = this.grid[i][j];
+          if(!cell.parent && !(cell.occupied || cell.hasTower)&& cell!=this.root){
+            return false;
 
+          }
+        }
+      }
+      return true;
+    }
+    //undo an invalid map action
+    undo(cell,tower) {
+      if(tower){
+        return function() {
+          cell.hasTower=false;
+          towerGame.towers.splice(towerGame.towers.indexOf(tower))
+          alert("you cannot place a tower here")
+        }
+      }else{
+        return function() {
+          cell.occupied= !cell.occupied
+          alert("performing that action would create an invalid grid")
+        }
+      }
+    }
     // sendEnemies()
     // Send a random number of enemies, up to 5, each from a random location
     // in the top half of the grid.  About half of the enemies will take the
@@ -211,7 +247,7 @@ class Game {
       for(let i = this.enemies.length-1; i >= 0; i--) {
         if(this.enemies[i].kill)
             this.enemies.splice(i,1);   // delete this dead enemy
-        else this.enemies[i].run();
+        
         }
     }
 
@@ -268,9 +304,7 @@ class Game {
       this.grid[i] = [];
       for(var j = 0; j < this.rows; j++){
         this.grid[i][j] = new Cell(this, vector2d((i*this.w), (j*this.w)), ++cellId);
-        // make 10% of the cells occupied
-        if(this.grid[i][j] != this.root && Math.floor(Math.random()*100) < 10)
-            this.grid[i][j].occupied = true;
+
       }
     }
 
@@ -326,10 +360,11 @@ class Game {
     // Some money required but also cannot place tower on a cell
     // of the grid that is occupied or is the root cell
     if(towerGame.placingTower) {
-        if(!cell.occupied && !cell.hasTower && cell != towerGame.root)
-            return true;
-      }
-    return(false);
+        if(!cell.occupied && !cell.hasTower && cell != towerGame.root){
+          return true;
+        }
+      return(false);
+    }
   }
 
   createTower(mtd) { // menu turret div
@@ -359,7 +394,7 @@ class Game {
     // placing a tower makes the cell containing the tower
     // unavailable to enemies the same as if it were
     // occupied (blocked)
-    towerGame.brushfire();   // all new distances and parents
+    towerGame.brushfire(towerGame.undo(cell,towerGame.towers[towerGame.towers.length-1]));   // all new distances and parents
   }
 
 //+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++ load callbacks
@@ -428,7 +463,7 @@ class Game {
     else if(!towerGame.placingTower && !cell.hasTower) {
         // toggle the occupied property of the clicked cell
         cell.occupied = !cell.occupied;
-        towerGame.brushfire();   // all new distances and parents
+        towerGame.brushfire(towerGame.undo(cell));   // all new distances and parents
         }
   }
   //+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++ Other
